[#245] fix/이미지 로직을 이미지 도메인으로 이동
