+++ src/main/java/com/anipick/backend/image/service/ImageService.java
@@ -5,6 +5,7 @@ import com.anipick.backend.common.exception.CustomException;
 import com.anipick.backend.common.exception.ErrorCode;
 import com.anipick.backend.image.domain.Image;
 import com.anipick.backend.image.domain.ImageDefaults;
+import com.anipick.backend.image.dto.ImageIdResponse;
 import com.anipick.backend.image.mapper.ImageMapper;
 import lombok.RequiredArgsConstructor;
 import net.coobird.thumbnailator.Thumbnails;
@@ -14,6 +15,7 @@ import org.springframework.core.io.ClassPathResource;
 import org.springframework.core.io.FileSystemResource;
 import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.imageio.ImageIO;
@@ -84,7 +86,7 @@ public class ImageService {
     }
 
     public String getImageUrlEndpoint(Long imageId) {
-        return ImageDefaults.HTTP_IMAGE_ENDPOINT + imageId;
+        return ImageDefaults.IMAGE_ENDPOINT + imageId;
     }
 
     public Optional<Image> getImageByImageId(Long imageId) {
@@ -93,6 +95,22 @@ public class ImageService {
 
     public Optional<Image> getImageByAuthId(Long userId) {
         return imageMapper.findByUserId(userId);
+    }
+
+    @Transactional
+    public ImageIdResponse updateProfileImage(CustomUserDetails user, MultipartFile profileImageFile) {
+        String originalFilename = profileImageFile.getOriginalFilename();
+        Image image;
+
+        try {
+            File outputFile = compressAndSaveImageToServer(user, profileImageFile);
+            image = insertImage(user, originalFilename, outputFile);
+
+        } catch (IOException e) {
+            throw new CustomException(ErrorCode.INTERNAL_SERVER_ERROR);
+        }
+
+        return ImageIdResponse.from(image.getImageId());
     }
 
     public void deleteImage(Long imageId) {
