+++ src/main/java/com/anipick/backend/mypage/controller/MyPageController.java
@@ -2,22 +2,14 @@ package com.anipick.backend.mypage.controller;
 
 import com.anipick.backend.common.auth.dto.CustomUserDetails;
 import com.anipick.backend.common.dto.ApiResponse;
-import com.anipick.backend.common.exception.CustomException;
-import com.anipick.backend.common.exception.ErrorCode;
 import com.anipick.backend.mypage.dto.*;
 import com.anipick.backend.mypage.service.MyPageService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.core.io.Resource;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
+
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
-import java.nio.file.Files;
 
 @RestController
 @RequiredArgsConstructor
@@ -32,33 +24,6 @@ public class MyPageController {
         MyPageResponse response = myPageService.getMyPage(userId);
         return ApiResponse.success(response);
     }
-
-    @PostMapping("/profile-image")
-    public ApiResponse<ImageIdResponse> updateProfileImage(
-            @AuthenticationPrincipal CustomUserDetails user,
-            @RequestPart("profileImageFile") MultipartFile profileImageFile
-    ) {
-        ImageIdResponse response = myPageService.updateProfileImage(user, profileImageFile);
-        return ApiResponse.success(response);
-    }
-
-//    @GetMapping("/profile-image/{imageId}")
-//    public ResponseEntity<Resource> getProfileImage(
-//            @AuthenticationPrincipal CustomUserDetails user,
-//            @PathVariable("imageId") Long imageId
-//    ) {
-//        try {
-//            Resource resource = myPageService.getProfileImage(user, imageId);
-//            String contentType = Files.probeContentType(resource.getFile().toPath());
-//            return ResponseEntity.ok()
-//                    .contentType(MediaType.parseMediaType(contentType))
-//                    .body(resource);
-//        } catch (IOException e) {
-//            log.error("IO error : ", e);
-//        }
-//
-//        return ResponseEntity.notFound().build();
-//    }
 
     @GetMapping("/animes/watchlist")
     public ApiResponse<WatchListAnimesResponse> getMyAnimesWatchList(
